Refactor: extract entry image widget
