+++ lotti/lib/widgets/journal/entry_detail_widget.dart
@@ -9,11 +9,11 @@ import 'package:lotti/classes/journal_entities.dart';
 import 'package:lotti/database/database.dart';
 import 'package:lotti/main.dart';
 import 'package:lotti/theme.dart';
-import 'package:lotti/utils/image_utils.dart';
 import 'package:lotti/widgets/audio/audio_player.dart';
 import 'package:lotti/widgets/journal/editor_tools.dart';
 import 'package:lotti/widgets/journal/editor_widget.dart';
 import 'package:lotti/widgets/journal/entry_datetime_modal.dart';
+import 'package:lotti/widgets/journal/entry_image_widget.dart';
 import 'package:lotti/widgets/journal/entry_tools.dart';
 import 'package:lotti/widgets/misc/map_widget.dart';
 import 'package:lotti/widgets/misc/survey_summary.dart';
@@ -236,58 +236,5 @@ class _EntryDetailWidgetState extends State<EntryDetailWidget> {
         );
       },
     );
-  }
-}
-
-class EntryImageWidget extends StatefulWidget {
-  final JournalImage journalImage;
-  final int height;
-  final BoxFit fit;
-
-  const EntryImageWidget(
-      {Key? key,
-      required this.journalImage,
-      required this.height,
-      this.fit = BoxFit.scaleDown})
-      : super(key: key);
-
-  @override
-  State<EntryImageWidget> createState() => _EntryImageWidgetState();
-}
-
-class _EntryImageWidgetState extends State<EntryImageWidget> {
-  Directory? docDir;
-
-  @override
-  void initState() {
-    super.initState();
-
-    getApplicationDocumentsDirectory().then((value) {
-      setState(() {
-        docDir = value;
-      });
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    if (docDir != null) {
-      File file =
-          File(getFullImagePathWithDocDir(widget.journalImage, docDir!));
-
-      return Container(
-        color: Colors.black,
-        height: widget.height.toDouble(),
-        child: Image.file(
-          file,
-          cacheHeight: widget.height * 3,
-          width: double.infinity,
-          height: widget.height.toDouble(),
-          fit: widget.fit,
-        ),
-      );
-    } else {
-      return Container();
-    }
   }
 }
